pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "arca-bot"
version = "1.3.7"
description = "Bot Discord multipropósito para a organização ARCA (Star Citizen)"
authors = [
    {name = "ARCA Organization", email = "contato@arca-org.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "discord.py>=2.3.0",
    "python-dotenv>=1.0.0",
    "PyNaCl>=1.5.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "coverage>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0"
]

[project.urls]
Homepage = "https://github.com/SrGoes/arca-bot"
Repository = "https://github.com/SrGoes/arca-bot"
Issues = "https://github.com/SrGoes/arca-bot/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  __pycache__
  | \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | venv
  | build
  | dist
)/
'''

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError"
]

# Estrutura do projeto
[tool.project_structure]
main = "src/main.py"
core = "src/core/"
modules = "src/modules/"
commands = "src/commands/"

# Scripts do projeto
[tool.scripts]
run = "python run.py"
test = "python -m pytest tests/"
lint = "python -m flake8 src/"
